Refactoring

+++ modules/landing_zone/scripts/showoutput.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 
+# Exit if any of the intermediate steps fail
+set -e
+
 ## Source path
 OUTPUT_FILE_PATH=${1}
 if [ -z "${OUTPUT_FILE_PATH}" ]; then
@@ -7,5 +10,5 @@ if [ -z "${OUTPUT_FILE_PATH}" ]; then
   exit 1
 fi
 
-cat ${OUTPUT_FILE_PATH}
-rm ${OUTPUT_FILE_PATH}
+if [ -f ${OUTPUT_FILE_PATH} ]; then cat ${OUTPUT_FILE_PATH}; fi
+if [ -f ${OUTPUT_FILE_PATH} ]; then rm ${OUTPUT_FILE_PATH}; fi
